qc/modules: relocate pipeline tests for QC modules.
