+++ auto_process_ngs/test/qc/modules/test_qualimap_rnaseq.py
@@ -1,24 +1,45 @@
 #######################################################################
-# Unit tests for qc/pipeline.py ('rseqc_genebody_coverage' QC module)
+# Unit tests for qc/pipeline.py ('qualimap_rnaseq' QC module)
 #######################################################################
 
-# All imports declared in __init__.py file
-from . import *
+import unittest
+import tempfile
+import shutil
+import os
+from bcftbx.JobRunner import SimpleJobRunner
+from auto_process_ngs.metadata import AnalysisProjectQCDirInfo
+from auto_process_ngs.mock import MockGtf2bed
+from auto_process_ngs.mock import MockQualimap
+from auto_process_ngs.mock import MockRSeQC
+from auto_process_ngs.mock import MockSamtools
+from auto_process_ngs.mock import MockStar
+from auto_process_ngs.mock import MockAnalysisProject
+from auto_process_ngs.mock import UpdateAnalysisProject
+from auto_process_ngs.analysis import AnalysisProject
+from auto_process_ngs.qc.protocols import QCProtocol
+from auto_process_ngs.qc.pipeline import QCPipeline
+from ..protocols import BaseQCPipelineTestCase
 
-class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
+# Set to False to keep test output dirs
+REMOVE_TEST_OUTPUTS = True
+
+# Polling interval for pipeline
+POLL_INTERVAL = 0.1
+
+class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
     """
-    Tests for 'rseqc_genebody_coverage' QC module
+    Tests for 'qualimap_rnaseq' QC module
     """
-    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_pe(self):
+    def test_qcpipeline_qc_modules_qualimap_rnaseq_pe(self):
         """
-        QCPipeline: 'rseqc_genebody_coverage' QC module (PE data)
+        QCPipeline: 'qualimap_rnaseq' QC module (PE data)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
+        MockQualimap.create(os.path.join(self.bin,"qualimap"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -29,38 +50,36 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="rseqc_genebody_coverage",
-                              description="RSeqc_genebody_coverage test",
+        protocol = QCProtocol(name="qualimap_rnaseq",
+                              description="Qualimap_rnaseq test",
                               seq_data_reads=['r1','r2'],
                               index_reads=None,
-                              qc_modules=("rseqc_genebody_coverage",))
+                              qc_modules=("qualimap_rnaseq",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
                           protocol)
         status = runqc.run(star_indexes=
                            { 'human': '/data/hg38/star_index' },
-                           annotation_bed_files=
-                           { 'human': self.ref_data['hg38']['bed'] },
+                           annotation_gtf_files=
+                           { 'human': self.ref_data['hg38']['gtf'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
-                                                   "rseqc_genebody_coverage",
-                                                   "human")
-        for f in ("PJB.geneBodyCoverage.curves.png",
-                  "PJB.geneBodyCoverage.r",
-                  "PJB.geneBodyCoverage.txt"):
-            self.assertTrue(os.path.exists(
-                os.path.join(rseqc_genebody_coverage_dir,f)),
+        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
+        for f in ("PJB1_S1_001/qualimapReport.html",
+                  "PJB1_S1_001/rnaseq_qc_results.txt",
+                  "PJB2_S2_001/qualimapReport.html",
+                  "PJB2_S2_001/rnaseq_qc_results.txt"):
+            self.assertTrue(os.path.exists(os.path.join(qualimap_dir,f)),
                             "%s: missing" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
+        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -75,8 +94,8 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,False)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
-        self.assertEqual(qc_info.annotation_gtf,None)
+        self.assertEqual(qc_info.annotation_bed,None)
+        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -87,16 +106,16 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_pe_with_r1_index_reads(self):
+    def test_qcpipeline_qc_modules_qualimap_rnaseq_pe_with_r1_index_reads(self):
         """
-        QCPipeline: 'rseqc_genebody_coverage' QC module (PE data, R1 as index reads)
+        QCPipeline: 'qualimap_rnaseq' QC module (PE data, R1 as index reads)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
+        MockQualimap.create(os.path.join(self.bin,"qualimap"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -107,38 +126,36 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="rseqc_genebody_coverage",
-                              description="RSeqc_genebody_coverage test",
+        protocol = QCProtocol(name="qualimap_rnaseq",
+                              description="Qualimap_rnaseq test",
                               seq_data_reads=['r2'],
                               index_reads=['r1'],
-                              qc_modules=("rseqc_genebody_coverage",))
+                              qc_modules=("qualimap_rnaseq",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
                           protocol)
         status = runqc.run(star_indexes=
                            { 'human': '/data/hg38/star_index' },
-                           annotation_bed_files=
-                           { 'human': self.ref_data['hg38']['bed'] },
+                           annotation_gtf_files=
+                           { 'human': self.ref_data['hg38']['gtf'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
-                                                   "rseqc_genebody_coverage",
-                                                   "human")
-        for f in ("PJB.geneBodyCoverage.curves.png",
-                  "PJB.geneBodyCoverage.r",
-                  "PJB.geneBodyCoverage.txt"):
-            self.assertTrue(os.path.exists(
-                os.path.join(rseqc_genebody_coverage_dir,f)),
+        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
+        for f in ("PJB1_S1_001/qualimapReport.html",
+                  "PJB1_S1_001/rnaseq_qc_results.txt",
+                  "PJB2_S2_001/qualimapReport.html",
+                  "PJB2_S2_001/rnaseq_qc_results.txt"):
+            self.assertTrue(os.path.exists(os.path.join(qualimap_dir,f)),
                             "%s: missing" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
+        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -153,8 +170,8 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,False)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
-        self.assertEqual(qc_info.annotation_gtf,None)
+        self.assertEqual(qc_info.annotation_bed,None)
+        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -165,16 +182,16 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_se(self):
+    def test_qcpipeline_qc_modules_qualimap_rnaseq_se(self):
         """
-        QCPipeline: 'rseqc_genebody_coverage' QC module (SE data)
+        QCPipeline: 'qualimap_rnaseq' QC module (SE data)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
+        MockQualimap.create(os.path.join(self.bin,"qualimap"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -183,38 +200,36 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="rseqc_genebody_coverage",
-                              description="RSeqc_genebody_coverage test",
-                              seq_data_reads=['r1',],
+        protocol = QCProtocol(name="qualimap_rnaseq",
+                              description="Qualimap_rnaseq test",
+                              seq_data_reads=['r1'],
                               index_reads=None,
-                              qc_modules=("rseqc_genebody_coverage",))
+                              qc_modules=("qualimap_rnaseq",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
                           protocol)
         status = runqc.run(star_indexes=
                            { 'human': '/data/hg38/star_index' },
-                           annotation_bed_files=
-                           { 'human': self.ref_data['hg38']['bed'] },
+                           annotation_gtf_files=
+                           { 'human': self.ref_data['hg38']['gtf'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
-                                                   "rseqc_genebody_coverage",
-                                                   "human")
-        for f in ("PJB.geneBodyCoverage.curves.png",
-                  "PJB.geneBodyCoverage.r",
-                  "PJB.geneBodyCoverage.txt"):
-            self.assertTrue(os.path.exists(
-                os.path.join(rseqc_genebody_coverage_dir,f)),
+        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
+        for f in ("PJB1_S1_001/qualimapReport.html",
+                  "PJB1_S1_001/rnaseq_qc_results.txt",
+                  "PJB2_S2_001/qualimapReport.html",
+                  "PJB2_S2_001/rnaseq_qc_results.txt"):
+            self.assertTrue(os.path.exists(os.path.join(qualimap_dir,f)),
                             "%s: missing" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
+        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -227,8 +242,8 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,False)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
-        self.assertEqual(qc_info.annotation_gtf,None)
+        self.assertEqual(qc_info.annotation_bed,None)
+        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -239,16 +254,16 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_se_with_biological_samples(self):
+    def test_qcpipeline_qc_modules_qualimap_rnaseq_se_with_biological_samples(self):
         """
-        QCPipeline: 'rseqc_genebody_coverage' QC module (SE data with biological samples)
+        QCPipeline: 'qualimap_rnaseq' QC module (SE data with biological samples)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
+        MockQualimap.create(os.path.join(self.bin,"qualimap"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -258,38 +273,39 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                            'Biological samples': 'PJB1' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="rseqc_genebody_coverage",
-                              description="RSeqc_genebody_coverage test",
-                              seq_data_reads=['r1',],
+        protocol = QCProtocol(name="qualimap_rnaseq",
+                              description="Qualimap_rnaseq test",
+                              seq_data_reads=['r1'],
                               index_reads=None,
-                              qc_modules=("rseqc_genebody_coverage",))
+                              qc_modules=("qualimap_rnaseq",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
                           protocol)
         status = runqc.run(star_indexes=
                            { 'human': '/data/hg38/star_index' },
-                           annotation_bed_files=
-                           { 'human': self.ref_data['hg38']['bed'] },
+                           annotation_gtf_files=
+                           { 'human': self.ref_data['hg38']['gtf'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
-                                                   "rseqc_genebody_coverage",
-                                                   "human")
-        for f in ("PJB.geneBodyCoverage.curves.png",
-                  "PJB.geneBodyCoverage.r",
-                  "PJB.geneBodyCoverage.txt"):
-            self.assertTrue(os.path.exists(
-                os.path.join(rseqc_genebody_coverage_dir,f)),
+        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
+        for f in ("PJB1_S1_001/qualimapReport.html",
+                  "PJB1_S1_001/rnaseq_qc_results.txt"):
+            self.assertTrue(os.path.exists(os.path.join(qualimap_dir,f)),
                             "%s: missing" % f)
+        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
+        for f in ("PJB2_S2_001/qualimapReport.html",
+                  "PJB2_S2_001/rnaseq_qc_results.txt"):
+            self.assertFalse(os.path.exists(os.path.join(qualimap_dir,f)),
+                            "%s: present" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
+        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -302,8 +318,8 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,False)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
-        self.assertEqual(qc_info.annotation_gtf,None)
+        self.assertEqual(qc_info.annotation_bed,None)
+        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -314,16 +330,16 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_se_split_lanes(self):
+    def test_qcpipeline_qc_modules_qualimap_rnaseq_se_split_lanes(self):
         """
-        QCPipeline: 'rseqc_genebody_coverage' QC module (SE data, split by lane)
+        QCPipeline: 'qualimap_rnaseq' QC module (SE data, split by lane)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
+        MockQualimap.create(os.path.join(self.bin,"qualimap"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -332,11 +348,11 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="rseqc_genebody_coverage",
-                              description="RSeqc_genebody_coverage test",
-                              seq_data_reads=['r1',],
+        protocol = QCProtocol(name="qualimap_rnaseq",
+                              description="Qualimap_rnaseq test",
+                              seq_data_reads=['r1'],
                               index_reads=None,
-                              qc_modules=("rseqc_genebody_coverage",))
+                              qc_modules=("qualimap_rnaseq",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
@@ -344,27 +360,25 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                           split_fastqs_by_lane=True)
         status = runqc.run(star_indexes=
                            { 'human': '/data/hg38/star_index' },
-                           annotation_bed_files=
-                           { 'human': self.ref_data['hg38']['bed'] },
+                           annotation_gtf_files=
+                           { 'human': self.ref_data['hg38']['gtf'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
-                                                   "rseqc_genebody_coverage",
-                                                   "human")
-        for f in ("PJB.geneBodyCoverage.curves.png",
-                  "PJB.geneBodyCoverage.r",
-                  "PJB.geneBodyCoverage.txt"):
-            self.assertTrue(os.path.exists(
-                os.path.join(rseqc_genebody_coverage_dir,f)),
+        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
+        for f in ("PJB1_S1_L001_001/qualimapReport.html",
+                  "PJB1_S1_L001_001/rnaseq_qc_results.txt",
+                  "PJB2_S2_L001_001/qualimapReport.html",
+                  "PJB2_S2_L001_001/rnaseq_qc_results.txt"):
+            self.assertTrue(os.path.exists(os.path.join(qualimap_dir,f)),
                             "%s: missing" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
+        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -377,8 +391,8 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,True)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
-        self.assertEqual(qc_info.annotation_gtf,None)
+        self.assertEqual(qc_info.annotation_bed,None)
+        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -389,16 +403,16 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_missing_STAR_index(self):
+    def test_qcpipeline_qc_modules_qualimap_rnaseq_missing_star_index(self):
         """
-        QCPipeline: 'rseqc_genebody_coverage' QC module (no STAR index)
+        QCPipeline: 'qualimap_rnaseq' QC module (no STAR index)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
+        MockQualimap.create(os.path.join(self.bin,"qualimap"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -407,36 +421,35 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="rseqc_genebody_coverage",
-                              description="RSeqc_genebody_coverage test",
-                              seq_data_reads=['r1',],
+        protocol = QCProtocol(name="qualimap_rnaseq",
+                              description="Qualimap_rnaseq test",
+                              seq_data_reads=['r1'],
                               index_reads=None,
-                              qc_modules=("rseqc_genebody_coverage",))
+                              qc_modules=("qualimap_rnaseq",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
                           protocol)
-        status = runqc.run(annotation_bed_files=
-                           { 'human': self.ref_data['hg38']['bed'] },
+        status = runqc.run(annotation_gtf_files=
+                           { 'human': self.ref_data['hg38']['gtf'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
+        self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
-                                                   "rseqc_genebody_coverage",
-                                                   "human")
-        for f in ("PJB.geneBodyCoverage.curves.png",
-                  "PJB.geneBodyCoverage.r",
-                  "PJB.geneBodyCoverage.txt"):
-            self.assertFalse(os.path.exists(
-                os.path.join(rseqc_genebody_coverage_dir,f)),
+        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
+        for f in ("PJB1_S1_001/qualimapReport.html",
+                  "PJB1_S1_001/rnaseq_qc_results.txt",
+                  "PJB2_S2_001/qualimapReport.html",
+                  "PJB2_S2_001/rnaseq_qc_results.txt"):
+            self.assertFalse(os.path.exists(os.path.join(qualimap_dir,f)),
                             "%s: present" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
+        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -449,8 +462,8 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,False)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,None)
-        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
-        self.assertEqual(qc_info.annotation_gtf,None)
+        self.assertEqual(qc_info.annotation_bed,None)
+        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -461,16 +474,16 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_missing_BED_file(self):
+    def test_qcpipeline_qc_modules_qualimap_rnaseq_missing_gtf_file(self):
         """
-        QCPipeline: 'rseqc_genebody_coverage' QC module (no BED file)
+        QCPipeline: 'qualimap_rnaseq' QC module (no GTF file)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
+        MockQualimap.create(os.path.join(self.bin,"qualimap"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -479,11 +492,11 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="rseqc_genebody_coverage",
-                              description="RSeqc_genebody_coverage test",
-                              seq_data_reads=['r1',],
+        protocol = QCProtocol(name="qualimap_rnaseq",
+                              description="Qualimap_rnaseq test",
+                              seq_data_reads=['r1'],
                               index_reads=None,
-                              qc_modules=("rseqc_genebody_coverage",))
+                              qc_modules=("qualimap_rnaseq",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
@@ -494,21 +507,20 @@ class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
+        self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
-                                                   "rseqc_genebody_coverage",
-                                                   "human")
-        for f in ("PJB.geneBodyCoverage.curves.png",
-                  "PJB.geneBodyCoverage.r",
-                  "PJB.geneBodyCoverage.txt"):
-            self.assertFalse(os.path.exists(
-                os.path.join(rseqc_genebody_coverage_dir,f)),
+        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
+        for f in ("PJB1_S1_001/qualimapReport.html",
+                  "PJB1_S1_001/rnaseq_qc_results.txt",
+                  "PJB2_S2_001/qualimapReport.html",
+                  "PJB2_S2_001/rnaseq_qc_results.txt"):
+            self.assertFalse(os.path.exists(os.path.join(qualimap_dir,f)),
                             "%s: present" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
+        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
