+++ auto_process_ngs/test/qc/modules/test_rseqc_genebody_coverage.py
@@ -1,24 +1,44 @@
 #######################################################################
-# Unit tests for qc/pipeline.py ('qualimap_rnaseq' QC module)
+# Unit tests for qc/pipeline.py ('rseqc_genebody_coverage' QC module)
 #######################################################################
 
-# All imports declared in __init__.py file
-from . import *
+import unittest
+import tempfile
+import shutil
+import os
+from bcftbx.JobRunner import SimpleJobRunner
+from auto_process_ngs.metadata import AnalysisProjectQCDirInfo
+from auto_process_ngs.mock import MockGtf2bed
+from auto_process_ngs.mock import MockRSeQC
+from auto_process_ngs.mock import MockSamtools
+from auto_process_ngs.mock import MockStar
+from auto_process_ngs.mock import MockAnalysisProject
+from auto_process_ngs.mock import UpdateAnalysisProject
+from auto_process_ngs.analysis import AnalysisProject
+from auto_process_ngs.qc.protocols import QCProtocol
+from auto_process_ngs.qc.pipeline import QCPipeline
+from ..protocols import BaseQCPipelineTestCase
 
-class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
+# Set to False to keep test output dirs
+REMOVE_TEST_OUTPUTS = True
+
+# Polling interval for pipeline
+POLL_INTERVAL = 0.1
+
+class TestQCPipelineRseqcGenebodyCoverage(BaseQCPipelineTestCase):
     """
-    Tests for 'qualimap_rnaseq' QC module
+    Tests for 'rseqc_genebody_coverage' QC module
     """
-    def test_qcpipeline_qc_modules_qualimap_rnaseq_pe(self):
+    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_pe(self):
         """
-        QCPipeline: 'qualimap_rnaseq' QC module (PE data)
+        QCPipeline: 'rseqc_genebody_coverage' QC module (PE data)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockQualimap.create(os.path.join(self.bin,"qualimap"))
+        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -29,36 +49,38 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="qualimap_rnaseq",
-                              description="Qualimap_rnaseq test",
+        protocol = QCProtocol(name="rseqc_genebody_coverage",
+                              description="RSeqc_genebody_coverage test",
                               seq_data_reads=['r1','r2'],
                               index_reads=None,
-                              qc_modules=("qualimap_rnaseq",))
+                              qc_modules=("rseqc_genebody_coverage",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
                           protocol)
         status = runqc.run(star_indexes=
                            { 'human': '/data/hg38/star_index' },
-                           annotation_gtf_files=
-                           { 'human': self.ref_data['hg38']['gtf'] },
+                           annotation_bed_files=
+                           { 'human': self.ref_data['hg38']['bed'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
-        for f in ("PJB1_S1_001/qualimapReport.html",
-                  "PJB1_S1_001/rnaseq_qc_results.txt",
-                  "PJB2_S2_001/qualimapReport.html",
-                  "PJB2_S2_001/rnaseq_qc_results.txt"):
-            self.assertTrue(os.path.exists(os.path.join(qualimap_dir,f)),
+        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
+                                                   "rseqc_genebody_coverage",
+                                                   "human")
+        for f in ("PJB.geneBodyCoverage.curves.png",
+                  "PJB.geneBodyCoverage.r",
+                  "PJB.geneBodyCoverage.txt"):
+            self.assertTrue(os.path.exists(
+                os.path.join(rseqc_genebody_coverage_dir,f)),
                             "%s: missing" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
+        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -73,8 +95,8 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,False)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,None)
-        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
+        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
+        self.assertEqual(qc_info.annotation_gtf,None)
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -85,16 +107,16 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_qualimap_rnaseq_pe_with_r1_index_reads(self):
+    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_pe_with_r1_index_reads(self):
         """
-        QCPipeline: 'qualimap_rnaseq' QC module (PE data, R1 as index reads)
+        QCPipeline: 'rseqc_genebody_coverage' QC module (PE data, R1 as index reads)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockQualimap.create(os.path.join(self.bin,"qualimap"))
+        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -105,36 +127,38 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="qualimap_rnaseq",
-                              description="Qualimap_rnaseq test",
+        protocol = QCProtocol(name="rseqc_genebody_coverage",
+                              description="RSeqc_genebody_coverage test",
                               seq_data_reads=['r2'],
                               index_reads=['r1'],
-                              qc_modules=("qualimap_rnaseq",))
+                              qc_modules=("rseqc_genebody_coverage",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
                           protocol)
         status = runqc.run(star_indexes=
                            { 'human': '/data/hg38/star_index' },
-                           annotation_gtf_files=
-                           { 'human': self.ref_data['hg38']['gtf'] },
+                           annotation_bed_files=
+                           { 'human': self.ref_data['hg38']['bed'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
-        for f in ("PJB1_S1_001/qualimapReport.html",
-                  "PJB1_S1_001/rnaseq_qc_results.txt",
-                  "PJB2_S2_001/qualimapReport.html",
-                  "PJB2_S2_001/rnaseq_qc_results.txt"):
-            self.assertTrue(os.path.exists(os.path.join(qualimap_dir,f)),
+        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
+                                                   "rseqc_genebody_coverage",
+                                                   "human")
+        for f in ("PJB.geneBodyCoverage.curves.png",
+                  "PJB.geneBodyCoverage.r",
+                  "PJB.geneBodyCoverage.txt"):
+            self.assertTrue(os.path.exists(
+                os.path.join(rseqc_genebody_coverage_dir,f)),
                             "%s: missing" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
+        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -149,8 +173,8 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,False)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,None)
-        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
+        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
+        self.assertEqual(qc_info.annotation_gtf,None)
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -161,16 +185,16 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_qualimap_rnaseq_se(self):
+    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_se(self):
         """
-        QCPipeline: 'qualimap_rnaseq' QC module (SE data)
+        QCPipeline: 'rseqc_genebody_coverage' QC module (SE data)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockQualimap.create(os.path.join(self.bin,"qualimap"))
+        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -179,36 +203,38 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="qualimap_rnaseq",
-                              description="Qualimap_rnaseq test",
-                              seq_data_reads=['r1'],
+        protocol = QCProtocol(name="rseqc_genebody_coverage",
+                              description="RSeqc_genebody_coverage test",
+                              seq_data_reads=['r1',],
                               index_reads=None,
-                              qc_modules=("qualimap_rnaseq",))
+                              qc_modules=("rseqc_genebody_coverage",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
                           protocol)
         status = runqc.run(star_indexes=
                            { 'human': '/data/hg38/star_index' },
-                           annotation_gtf_files=
-                           { 'human': self.ref_data['hg38']['gtf'] },
+                           annotation_bed_files=
+                           { 'human': self.ref_data['hg38']['bed'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
-        for f in ("PJB1_S1_001/qualimapReport.html",
-                  "PJB1_S1_001/rnaseq_qc_results.txt",
-                  "PJB2_S2_001/qualimapReport.html",
-                  "PJB2_S2_001/rnaseq_qc_results.txt"):
-            self.assertTrue(os.path.exists(os.path.join(qualimap_dir,f)),
+        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
+                                                   "rseqc_genebody_coverage",
+                                                   "human")
+        for f in ("PJB.geneBodyCoverage.curves.png",
+                  "PJB.geneBodyCoverage.r",
+                  "PJB.geneBodyCoverage.txt"):
+            self.assertTrue(os.path.exists(
+                os.path.join(rseqc_genebody_coverage_dir,f)),
                             "%s: missing" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
+        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -221,8 +247,8 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,False)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,None)
-        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
+        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
+        self.assertEqual(qc_info.annotation_gtf,None)
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -233,16 +259,16 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_qualimap_rnaseq_se_with_biological_samples(self):
+    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_se_with_biological_samples(self):
         """
-        QCPipeline: 'qualimap_rnaseq' QC module (SE data with biological samples)
+        QCPipeline: 'rseqc_genebody_coverage' QC module (SE data with biological samples)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockQualimap.create(os.path.join(self.bin,"qualimap"))
+        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -252,39 +278,38 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                            'Biological samples': 'PJB1' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="qualimap_rnaseq",
-                              description="Qualimap_rnaseq test",
-                              seq_data_reads=['r1'],
+        protocol = QCProtocol(name="rseqc_genebody_coverage",
+                              description="RSeqc_genebody_coverage test",
+                              seq_data_reads=['r1',],
                               index_reads=None,
-                              qc_modules=("qualimap_rnaseq",))
+                              qc_modules=("rseqc_genebody_coverage",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
                           protocol)
         status = runqc.run(star_indexes=
                            { 'human': '/data/hg38/star_index' },
-                           annotation_gtf_files=
-                           { 'human': self.ref_data['hg38']['gtf'] },
+                           annotation_bed_files=
+                           { 'human': self.ref_data['hg38']['bed'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
-        for f in ("PJB1_S1_001/qualimapReport.html",
-                  "PJB1_S1_001/rnaseq_qc_results.txt"):
-            self.assertTrue(os.path.exists(os.path.join(qualimap_dir,f)),
+        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
+                                                   "rseqc_genebody_coverage",
+                                                   "human")
+        for f in ("PJB.geneBodyCoverage.curves.png",
+                  "PJB.geneBodyCoverage.r",
+                  "PJB.geneBodyCoverage.txt"):
+            self.assertTrue(os.path.exists(
+                os.path.join(rseqc_genebody_coverage_dir,f)),
                             "%s: missing" % f)
-        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
-        for f in ("PJB2_S2_001/qualimapReport.html",
-                  "PJB2_S2_001/rnaseq_qc_results.txt"):
-            self.assertFalse(os.path.exists(os.path.join(qualimap_dir,f)),
-                            "%s: present" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
+        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -297,8 +322,8 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,False)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,None)
-        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
+        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
+        self.assertEqual(qc_info.annotation_gtf,None)
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -309,16 +334,16 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_qualimap_rnaseq_se_split_lanes(self):
+    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_se_split_lanes(self):
         """
-        QCPipeline: 'qualimap_rnaseq' QC module (SE data, split by lane)
+        QCPipeline: 'rseqc_genebody_coverage' QC module (SE data, split by lane)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockQualimap.create(os.path.join(self.bin,"qualimap"))
+        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -327,11 +352,11 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="qualimap_rnaseq",
-                              description="Qualimap_rnaseq test",
-                              seq_data_reads=['r1'],
+        protocol = QCProtocol(name="rseqc_genebody_coverage",
+                              description="RSeqc_genebody_coverage test",
+                              seq_data_reads=['r1',],
                               index_reads=None,
-                              qc_modules=("qualimap_rnaseq",))
+                              qc_modules=("rseqc_genebody_coverage",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
@@ -339,25 +364,27 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                           split_fastqs_by_lane=True)
         status = runqc.run(star_indexes=
                            { 'human': '/data/hg38/star_index' },
-                           annotation_gtf_files=
-                           { 'human': self.ref_data['hg38']['gtf'] },
+                           annotation_bed_files=
+                           { 'human': self.ref_data['hg38']['bed'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
-        for f in ("PJB1_S1_L001_001/qualimapReport.html",
-                  "PJB1_S1_L001_001/rnaseq_qc_results.txt",
-                  "PJB2_S2_L001_001/qualimapReport.html",
-                  "PJB2_S2_L001_001/rnaseq_qc_results.txt"):
-            self.assertTrue(os.path.exists(os.path.join(qualimap_dir,f)),
+        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
+                                                   "rseqc_genebody_coverage",
+                                                   "human")
+        for f in ("PJB.geneBodyCoverage.curves.png",
+                  "PJB.geneBodyCoverage.r",
+                  "PJB.geneBodyCoverage.txt"):
+            self.assertTrue(os.path.exists(
+                os.path.join(rseqc_genebody_coverage_dir,f)),
                             "%s: missing" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
+        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -370,8 +397,8 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,True)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,None)
-        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
+        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
+        self.assertEqual(qc_info.annotation_gtf,None)
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -382,16 +409,16 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_qualimap_rnaseq_missing_star_index(self):
+    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_missing_STAR_index(self):
         """
-        QCPipeline: 'qualimap_rnaseq' QC module (no STAR index)
+        QCPipeline: 'rseqc_genebody_coverage' QC module (no STAR index)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockQualimap.create(os.path.join(self.bin,"qualimap"))
+        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -400,35 +427,36 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="qualimap_rnaseq",
-                              description="Qualimap_rnaseq test",
-                              seq_data_reads=['r1'],
+        protocol = QCProtocol(name="rseqc_genebody_coverage",
+                              description="RSeqc_genebody_coverage test",
+                              seq_data_reads=['r1',],
                               index_reads=None,
-                              qc_modules=("qualimap_rnaseq",))
+                              qc_modules=("rseqc_genebody_coverage",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
                           protocol)
-        status = runqc.run(annotation_gtf_files=
-                           { 'human': self.ref_data['hg38']['gtf'] },
+        status = runqc.run(annotation_bed_files=
+                           { 'human': self.ref_data['hg38']['bed'] },
                            poll_interval=POLL_INTERVAL,
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
-        self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
-        for f in ("PJB1_S1_001/qualimapReport.html",
-                  "PJB1_S1_001/rnaseq_qc_results.txt",
-                  "PJB2_S2_001/qualimapReport.html",
-                  "PJB2_S2_001/rnaseq_qc_results.txt"):
-            self.assertFalse(os.path.exists(os.path.join(qualimap_dir,f)),
+        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
+                                                   "rseqc_genebody_coverage",
+                                                   "human")
+        for f in ("PJB.geneBodyCoverage.curves.png",
+                  "PJB.geneBodyCoverage.r",
+                  "PJB.geneBodyCoverage.txt"):
+            self.assertFalse(os.path.exists(
+                os.path.join(rseqc_genebody_coverage_dir,f)),
                             "%s: present" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
+        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
@@ -441,8 +469,8 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
         self.assertEqual(qc_info.fastqs_split_by_lane,False)
         self.assertEqual(qc_info.fastq_screens,None)
         self.assertEqual(qc_info.star_index,None)
-        self.assertEqual(qc_info.annotation_bed,None)
-        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
+        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
+        self.assertEqual(qc_info.annotation_gtf,None)
         self.assertEqual(qc_info.cellranger_version,None)
         self.assertEqual(qc_info.cellranger_refdata,None)
         self.assertEqual(qc_info.cellranger_probeset,None)
@@ -453,16 +481,16 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                                         "PJB",f)),
                             "Missing %s" % f)
 
-    def test_qcpipeline_qc_modules_qualimap_rnaseq_missing_gtf_file(self):
+    def test_qcpipeline_qc_modules_rseqc_genebody_coverage_missing_BED_file(self):
         """
-        QCPipeline: 'qualimap_rnaseq' QC module (no GTF file)
+        QCPipeline: 'rseqc_genebody_coverage' QC module (no BED file)
         """
         # Make mock QC executables
         MockStar.create(os.path.join(self.bin,"STAR"))
         MockSamtools.create(os.path.join(self.bin,"samtools"))
         MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
         MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockQualimap.create(os.path.join(self.bin,"qualimap"))
+        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
         os.environ['PATH'] = "%s:%s" % (self.bin,
                                         os.environ['PATH'])
         # Make mock analysis project
@@ -471,11 +499,11 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                                 metadata={ 'Organism': 'Human' })
         p.create(top_dir=self.wd)
         # QC protocol
-        protocol = QCProtocol(name="qualimap_rnaseq",
-                              description="Qualimap_rnaseq test",
-                              seq_data_reads=['r1'],
+        protocol = QCProtocol(name="rseqc_genebody_coverage",
+                              description="RSeqc_genebody_coverage test",
+                              seq_data_reads=['r1',],
                               index_reads=None,
-                              qc_modules=("qualimap_rnaseq",))
+                              qc_modules=("rseqc_genebody_coverage",))
         # Set up and run the QC
         runqc = QCPipeline()
         runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
@@ -486,20 +514,21 @@ class TestQCPipelineQualimapRnaseq(BaseQCPipelineTestCase):
                            max_jobs=1,
                            runners={ 'default': SimpleJobRunner(), })
         self.assertEqual(status,0)
-        self.assertEqual(status,0)
         # Check outputs
         qc_dir = os.path.join(self.wd,"PJB","qc")
-        qualimap_dir = os.path.join(qc_dir,"qualimap-rnaseq","human")
-        for f in ("PJB1_S1_001/qualimapReport.html",
-                  "PJB1_S1_001/rnaseq_qc_results.txt",
-                  "PJB2_S2_001/qualimapReport.html",
-                  "PJB2_S2_001/rnaseq_qc_results.txt"):
-            self.assertFalse(os.path.exists(os.path.join(qualimap_dir,f)),
+        rseqc_genebody_coverage_dir = os.path.join(qc_dir,
+                                                   "rseqc_genebody_coverage",
+                                                   "human")
+        for f in ("PJB.geneBodyCoverage.curves.png",
+                  "PJB.geneBodyCoverage.r",
+                  "PJB.geneBodyCoverage.txt"):
+            self.assertFalse(os.path.exists(
+                os.path.join(rseqc_genebody_coverage_dir,f)),
                             "%s: present" % f)
         # Check QC metadata
         qc_info = AnalysisProjectQCDirInfo(
             os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"qualimap_rnaseq")
+        self.assertEqual(qc_info.protocol,"rseqc_genebody_coverage")
         self.assertEqual(qc_info.protocol_specification,
                          str(protocol))
         self.assertEqual(qc_info.organism,"Human")
